RPS-1110: Reject bare/malformed Docker & Helm OCI push input instead of 500ing (#461)
Replaces every hand-built bare or plain-text error response in the Docker and
Helm OCI handlers with a thrown panel exception, so ErrorHandler/OciErrorBodyAdvice
render it as the standard OCI distribution errors[] body instead of a bodyless
status or a flat 500.

RPS-1110: missing Content-Type on a Docker/Helm manifest push, a missing digest
on a Docker/Helm blob-upload finalize, an unacceptable pull Accept header, and an
unsupported manifest Content-Type on push now all answer with the proper thrown
exception (BadRequestException / HttpMediaTypeNotAcceptableException) instead of
a bare ResponseEntity or an uncaught IllegalArgumentException. MediaTypes.
getPreferredMediaType now returns null (instead of silently defaulting) when an
explicit, non-wildcard Accept header names nothing this registry serves, which is
what makes the pull handler's 406 path reachable at all.

RPS-1151: DockerManifestValidator now rejects a schemaVersion that does not match
what its media type requires (1 for the legacy Docker schema1, 2 for everything
else), closing the truncate-to-int bug where an out-of-range value silently
became whatever fit. ManifestTxService's two remaining (int) casts are replaced
with Math.toIntExact as a backstop.

RPS-1116: AbstractDockerProtocolTxFacade.extractPlatform now distinguishes an
image-config media type (DOCKER_CONFIG_JSON/OCI_CONFIG_JSON), which must be a
JSON object with os/architecture or the push is refused with 400
manifestConfigInvalid, from any other config media type (an OCI artifact), which
is stored under DockerConstants.UNKNOWN_PLATFORM without reading the blob at
all. Fixes the ordering bug where createManifest wrote the manifest to disk
before extracting the platform, so a rejected push left an orphaned manifest
file a later GET could still resolve; platform extraction now runs first.

RPS-1117: DockerManifestValidator.validateIndex no longer requires an index
entry to carry a platform (still rejects a null entry or a blank digest), and
findPlatformManifests stores a platform-less entry under the same
DockerConstants.UNKNOWN_PLATFORM constant RPS-1116 introduces.

Flips the e2e test.fail pins this batch closes (docker/registry-rules.spec.ts
R5 and R12/B5) and updates helm/registry-rules.spec.ts's R4 to expect the OCI
envelope instead of a bodyless 400.

Verification: mvn com.spotify.fmt:fmt-maven-plugin:format (0 reformatted) and
mvn verify -pl repsy-backend -am -Drat.skip=true are green. The e2e docker and
helm protocol suites were rebuilt and run twice each against a fresh local
stack, all green. A flip-and-fail check (temporarily reverting the
extractPlatform/writeManifest ordering fix) confirmed
imageConfigWithoutOsOrArchitectureIsRejected and nonJsonImageConfigIsRejected
catch the leaked manifest file, then the fix was restored.

RPS-1151, RPS-1116, RPS-1117

+++ repsy-backend/src/main/java/io/repsy/os/server/protocols/docker/shared/tag/services/ManifestTxService.java
@@ -393,7 +393,7 @@ public class ManifestTxService implements ManifestService<UUID> {
     manifest.setMediaType(newManifest.getMediaType());
     manifest.setConfigMediaType(newManifest.getConfig().getMediaType());
     manifest.setConfigDigest(newManifest.getConfigDigest());
-    manifest.setSchemaVersion((int) newManifest.getSchemaVersion());
+    manifest.setSchemaVersion(Math.toIntExact(newManifest.getSchemaVersion()));
     manifest.setLayers(layers);
     manifest.setPlatform(newManifest.getPlatform());
     manifest.setConfigSize(configLayer.getSize());
@@ -586,7 +586,7 @@ public class ManifestTxService implements ManifestService<UUID> {
     manifest.setMediaType(manifestInfo.getMediaType());
     manifest.setConfigMediaType(manifestInfo.getConfig().getMediaType());
     manifest.setConfigDigest(manifestInfo.getConfig().getDigest());
-    manifest.setSchemaVersion((int) manifestInfo.getSchemaVersion());
+    manifest.setSchemaVersion(Math.toIntExact(manifestInfo.getSchemaVersion()));
     manifest.setLayers(new HashSet<>());
     manifest.setPlatform(form.getPlatform());
